backend/requirements.txt
asgiref                       3.7.2
Brotli                        1.1.0
certifi                       2023.11.17
cffi                          1.16.0
charset-normalizer            3.3.2
cryptography                  41.0.7
cssselect2                    0.7.0
defusedxml                    0.8.0rc2
Django                        5.0
django-cors-headers           4.3.1
django-filter                 23.5
django-templated-mail         1.1.1
djangorestframework           3.14.0
djangorestframework-simplejwt 5.3.1
djoser                        2.2.2
fonttools                     4.47.0
html5lib                      1.1
idna                          3.6
Markdown                      3.5.1
mysqlclient                   2.2.1
oauthlib                      3.2.2
Pillow                        10.1.0
pip                           23.3.2
pycparser                     2.21
pydyf                         0.8.0
PyJWT                         2.8.0
pyphen                        0.14.0
python3-openid                3.2.0
pytz                          2023.3.post1
requests                      2.31.0
requests-oauthlib             1.3.1
six                           1.16.0
social-auth-app-django        5.4.0
social-auth-core              4.5.1
sqlparse                      0.4.4
tinycss2                      1.2.1
tzdata                        2023.3
urllib3                       2.1.0
weasyprint                    60.2
webencodings                  0.5.1
zopfli                        0.2.3